feat: Update sign in and sign up pages
Refactor the sign in and sign up pages to improve user experience and functionality. Update the layout, text, and input fields to enhance the overall design. Add a "Forgot password?" button to the sign in page. Implement a custom button component for the sign in and sign up pages.

Recent user commits:
- chore: Update input field colors and styles
- chore: Update skip and Get started buttons in Onboarding.dart and intro page 3 layout and text
- chore: Update intro page 1 layout and text

Recent repository commits:
- feat: Add Onboarding page to the travel app
- chore: Add intro page 3 to the travel app
- chore: Add intro page 2 to the travel app
- refactor: CustomInputField components, remove unused code and comments
- chore: Remove unused code and comments, refactor QrGenerator and CustomInputField components
- refactor: Replace SideMenu component with NavigationBar component
- chore: Add video player and device preview dependencies